chore: refactor
Refactor media package

+++ Lab04/AimsProject/src/hust/soict/hedspi/aims/cart/Cart.java
@@ -1,6 +1,5 @@
 package hust.soict.hedspi.aims.cart;
 
-import hust.soict.hedspi.aims.media.DigitalVideoDisc;
 import hust.soict.hedspi.aims.media.Media;
 
 import java.util.ArrayList;
@@ -64,8 +63,8 @@ public class Cart {
         System.out.println("***************************************************");
     }
 
-    public float totalCost() {
-        float cost = 0f;
+    public double totalCost() {
+        double cost = 0f;
         for (Media mediaInList : itemsOrdered) cost += mediaInList.getCost();
         return cost;
     }
